--- server/engine.js
@@ -1,50 +1,0 @@
-import { enqueueAddresses } from "./balanceQueue.js";
-import memory from "./memory.js";
-import logger from "./logger.js";
-
-const engine = async () => {
-  // Create a flat list of all addresses with their collection info
-  const allAddresses = [];
-  for (const [collectionName, collection] of Object.entries(
-    memory.db.collections
-  )) {
-    // Add regular addresses
-    collection.addresses.forEach((addr) => {
-      allAddresses.push({
-        ...addr,
-        collection: collectionName,
-      });
-    });
-
-    // Add extended key addresses
-    if (collection.extendedKeys) {
-      collection.extendedKeys.forEach((extendedKey) => {
-        extendedKey.addresses.forEach((addr) => {
-          allAddresses.push({
-            ...addr,
-            collection: collectionName,
-          });
-        });
-      });
-    }
-
-    // Add descriptor addresses
-    if (collection.descriptors) {
-      collection.descriptors.forEach((descriptor) => {
-        descriptor.addresses.forEach((addr) => {
-          allAddresses.push({
-            ...addr,
-            collection: collectionName,
-          });
-        });
-      });
-    }
-  }
-
-  logger.info(`Adding ${allAddresses.length} addresses to balance queue`);
-
-  // Add all addresses to the queue
-  enqueueAddresses(allAddresses);
-};
-
-export default engine;
